Showcase the flexibility of `git fetch -p`
The example for that exhibits the way to 'prune' remote branches that
were deleted upstream wasn't flexible as it relied on the command
defaulting to the upstream of the current branch. This might lead
the reader into overlooking the flexibility of the `git fetch`.

Show that the 'upstream' can be mentioned in the command thus show
casing the flexibility of `git fetch`.