Clarify that rebasing just re-writes history
Rebasing fast-forwards when the tip of the branch is a descendent of
the tip of the upstream. In other cases it re-writes the history. This
re-write is what actually leads the user to 'force' update the remote.

So, clarify that a user has to force update only when the history is
re-written regardless of whether the branch was fast-forwarded.